update nationals component structure
