+++ ubyssey/static/src/js/components/Nationals/index.js
@@ -1,0 +1,1 @@
+export {default as Nationals } from './Nationals.jsx';
