+++ ubyssey/static/src/js/components/Nationals/utils.js
@@ -1,0 +1,1 @@
+export const mapPath = '<title>Soccer Map</title><path d="M36.25,231.75s-2,9.75,2,11.25a29.42,29.42,0,0,1,1,6.38s0.5,2.63,2.38,2.5A48.73,48.73,0,0,1,48,255.25s1.38,2,4.75,0c0,0,3.63-2.1,5.25.89s0.38,0,.38,0,5.88,14.24,8.63,16.49a8,8,0,0,1,.88,3.75s0.13,13.75-1.62,17.88c0,0,.5,18.63,3.63,23.63,0,0,3.25,5.75-.62,10.5,0,0-3,6.63-3.87,8,0,0-4.25,3.5,2.38,0,0,0,1.75-1.25,1.5-2.75,0,0,1.05-4.25,1.9-1.5,0,0-1.4,1.63,1.6,1.38,0,0,1,1.26-3.12,2.63,0,0-5.12,2.25-3.62,4,0,0,1.13,1.5-4.75.88,0,0-5.5,2.88-4.62,6.25,0,0,.25,2.88,1.38,3.25,0,0-3.87,3.5-1.25,5.88l0.75,9.5s0,1.38,1-.87c0,0,2-3,1.88-4.37,0,0,.25-1.5,1.75-0.62a2.84,2.84,0,0,0,3.5,0s3-3,2.88-1.2c0,0-1.37,1,1.13,1.2,0,0-.25,1.75-3,3.5,0,0-2.37,3.38,1.88,3.63,0,0,4.21,0,4.36,3,0,0-1.52.5,0,1.5L71.5,374.5s0.5-5.75-1.87-5.75a4.38,4.38,0,0,1-2.87-.37s-4.12-5.75-4.5,1.13c0,0-.75,6.25,1.25,7,0,0,1.25,1.5-.87,1.38,0,0-1.75-.25-1.12,2.63,0,0-.87,2,0.13,2.88,0,0-3.87,2.13.25,2.5,0,0,1.5-.25,1.63-0.87,0,0,.75.13,1.13,2,0,0,2.09,2.25,8.23-4.62,0,0,2.11-.75,0,1.75,0,0-4.23,6.25-10.36,4.88,0,0-2.25.75,1,1.88,0,0-1.12,2.88,3,0,0,0,4.13-3.25,6.5-1.62,0,0,3.75,1.38-1.37,1.5L71,392.15a2.79,2.79,0,0,0-2.62-.53s-3.75,1-3.75,2.25c0,0-5.37-.75-4,5.13,0,0,2.63,1.25-.25,1.88,0,0-1.12.13-.37,2,0,0-3.12.25-1.87,3.38a1.77,1.77,0,0,0,1.5,2.25c2,0.38.38,0,.38,0s1.88,3.13,1.88,4.75c0,0-.62,1.5,2.63-1.25a10.32,10.32,0,0,0,5.25,1.66,2.38,2.38,0,0,0,1.63,3s0,1.75,2.25,1.63c0,0-1.25,2.75-5.62,1.13,0,0-4.75-2.25-4.12,0a5.56,5.56,0,0,0,5,2.38l0.75,1.88h1.75s3.88,1.88,6.38.75c0,0,4.75-4.87,4.63-7.75,0,0,.75-0.75.88,1.38,0,0-2,8.25-5,7.88a1.2,1.2,0,0,0,1,1.88h1.63v3.38s-3.62,2.75-2.12,4.75c0,0,.13,2.25,4.5,1.75,0,0-1.5,4.13.88,5.88,0,0,1,2.5,1.5,3.88h3.75s-1,2.5,2.75,2.38c0,0,1.38,1.63-1.25,1.63,0,0-2.25,0-1.12,2.25,0,0-.5,2,3.25,1.63v1.5S266.67,507,390,504c0,0-2.25-3.31,1.31-3.19,0,0,1.19-61.31-.69-63.44,0,0,8.21-3.71,18.54-19.71L443.67,374a5.5,5.5,0,0,0-4.08-2.17,9.43,9.43,0,0,0-5.83-3.89H432s-1.08-3.86-6.71-3.77c0,0-1.21-2.83-2.38-2.83,0,0-7.83,6.25-8.08,6.25,0,0-4.75-.67-4.58-1.67,0,0,2.83-6.83.17-9.67,0,0,1.33-5.33-.75-5.92,0,0-.83-7.08-3.92-8.33,0,0,.33-4.33-5.08-3.42,0,0-2.75.25-2.67-.33,0,0-1.25-1.08-1.5-.17a2.76,2.76,0,0,0,1.08,2.25s-0.83,1.67-.5,2.17c0,0-.42.58-1.67-1.17,0,0,.17-1.17-1.42-1.17,0,0-1.5-.33,0-1.25,0,0,.67-3.83-1.58-4.67,0,0,.33-3.25-0.92-3.33,0,0-.17-2.17.75-2.5l-1.08-1.92s0.17-2.5,1.67-2.67c0,0-.67-3.25-0.42-3.33S195.5,328.5,36.25,231.75Z" transform="translate(-27.37 -231.75)" style="fill:#b11f24"/><path d="M47.38,408.25a8.8,8.8,0,0,1,7.5,1.75s8.25,12.75,14,15.13c0,0,3.75,1,4.13,2,0,0,3.38-.12,2.38,2.38,0,0-1.5,1.38-.87,2.75a4.39,4.39,0,0,1-.62,2.75,11.55,11.55,0,0,0,2.25,2.25,1.79,1.79,0,0,1,0,2.38s2.81,1.19,0,3.56c0,0,2.69,3.88,4.69,4,0,0,.75,2,0,3.31,0,0,2.35.88,0.33,2.88,0,0,2.6,8.31-.52,9.69,0,0-1.75.75-1.37-1a3.27,3.27,0,0,1-4-1.25l-2.06-2,0.19-2.75L71.25,456a19.8,19.8,0,0,1-1.75-3.06s-3.81-1.62-.31-2.5l1.31-.31s1.69-2.12-1-2.37c0,0-4.37-.81-5.94-0.25,0,0-.44-3.94-2.94-3.94,0,0,0-2.37,1.63-2.5a1,1,0,0,0-.56-1.31s-2.67-3.52-4.44-3.63H57c-1.75.25-.75-0.94,0.19-1.31a7.93,7.93,0,0,1,1.94-1.87c0.63-.19.44-1.81-1.62-1.37,0,0-2.06-.5-2.12-3.81,0,0-.5-3.06-1.37-3.12,0,0-1.31-1.5-.87-2.44,0,0-2.87-2.25-3.87-2.25,0,0,.06-2.12-1-2.19,0,0-.69-1.44,2-1.31a25.15,25.15,0,0,1,2.88,2.25s2.06-.75.38-2c0,0-1.94-3.12-4.25-2.87,0,0-1.62-1-1.19-1.62Z" transform="translate(-27.37 -231.75)" style="fill:#b11f24"/><path d="M32.88,356.19a2.35,2.35,0,0,1,2.31.56s4.75-1.12,1.56,2.5a2.35,2.35,0,0,0,.81,1.5s1.25,2.38-1,3.06c0,0-1.81,2.25-.81,3.5,0,0,.38.38-2.06,0.63,0,0-1.06,3.87.25,4.5a1.35,1.35,0,0,1-.69,1.88l-0.06,1.75s0.31,3.75-3.37,4.31c0,0-.62-2.39,1.38-3.06,0,0,1.31-.75-0.31-1,0,0-1-2,.69-2,0,0,.32-1.56-1.84-1.37,0,0-.9-2.81,0-3.06,0,0,1.29-1.41-.4-1.64,0,0-3.2-3,.3-4,0,0-2.31-5.19.44-7.37,0,0,.46-1.44,2.89.13Z" transform="translate(-27.37 -231.75)" style="fill:#b11f24"/><path d="M32.95,338.56s-0.51-2.81-2.51,0a3.17,3.17,0,0,0,0,2.5s-3.81-.75-2.94.69L28.56,343s-0.56,3.06.19,4c0,0-2.75,1.44-.06,2.88,0,0,2.44,1.88,1.75,3.31,0,0,.25,1.19,2,0.38a3,3,0,0,1,2.19.75s2.63,1,3.56-.62c0,0,4.31-4.69,5.56-4.62,0,0,3.81.25,2.44-2.62a1.68,1.68,0,0,0-2.5.31s-1.25.63-2.25-1.12a5.48,5.48,0,0,1-1.25-1.56H37.25s0.56-3.75-1.06-3.69C36.19,340.38,35.46,337.31,32.95,338.56Z" transform="translate(-27.37 -231.75)" style="fill:#b11f24"/><path d="M443.67,374l-34.5,43.67s-10.92,16.21-18.54,19.71c0,0,1.13,44.38.69,63.44,0,0,4.19-.81,2.44-3.06,0,0-.25-0.87,1-1.75,0,0,0-3.75-1.12-4.62a2.09,2.09,0,0,1,1.88-.5s0.63-2,1.88.13c0,0-1.12,1,2.25,1.63a3.86,3.86,0,0,0,3.25,2.25s2.13,2.5-1.62,2.13c0,0-2.38-1.08-1.25-1.87,0,0-2.12-.87-2,0.38,0,0-3.37.75-1,2.25,0,0,2.63.5,7.13,1.75,0,0,1.13,1.25-1.5,1.31,0,0-2.87,1.19-2.87,2.19,0,0-3.62.5-3.5,1.75a23,23,0,0,1,4.63,4.25s2-1.12,3.38,0c0,0,2.63.5,2.88-1.37,0,0-1.62-2.12.75-2.25,0,0,1.63-.74,1.63-1.25,0,0,.25,2.5,1.13,2.5,0,0,1.25.88,1.25,1.5l0.88-.25s0.75-3,2.88.13c0,0-1.12,1.75,2.88,1.75,0,0,1,.25.88,2h0.88s-0.75,3.13,1.38,3.13c0,0,2.25-1,1.38-2.75,0,0,1.25-1.12,2.88.63,0,0,2.25,2.75,4.5,2.13a5.4,5.4,0,0,0,6.88-1.75s1-2,2.5.5c0,0,0,2.29,3.92,1.83l11.71-1s0.88-1-.75-1.37a2.37,2.37,0,0,1,2.25-3s4-3.75,3.5-.25c0,0-1.25,3.88,1.5,1.63,0,0,.63-3,0-4.75,0,0,2.3-2.5.28-3.62,0,0-.29-1.5,2,1.13a5.36,5.36,0,0,1-.68,4.75s0,2.38,2,.5c0,0,.5-4.5,5.75-3.87,0,0,4,.38,4-1.5,0,0,2.5-1.88-6.25-1,0,0-5-.5-2.25-2.87,0,0,2.5-.12,2.38,1.38,0,0,.38.75,1.63,0.25,0,0,4.13-1,6.38.88a2,2,0,0,0,1.88-.87,1.86,1.86,0,0,1,2.13,1.25s0.88,0.38,1.75-1.37a1.07,1.07,0,0,1,1.5,1s2,0.88,3.25.13c0,0,1.5,5.38,1.38,7.25,0,0-1.25,1.25,3.13.63,0,0,.38,2.88,2.75,3,0,0,4.5,1.25,7.5-.75,0,0,4.38.38,2.75,5.5,0,0,.75,3.88,4.13,4.25,0,0-1.75,6.63,3.63,6.88,0,0,1.13,1.63-.75,1.88,0,0-.83,2.13,1.09,2.38,0,0-2.46,2.38,1.79,2.13,0,0,2.63.13,2.75,1.5,0,0,1.63-.37,4.13,1.13,0,0,2-1.62,6.5,0,0,0,2.25,1.88,4.88-.75a2.39,2.39,0,0,1,2.38.5s5.5,0.38,7.63-.87c0,0,3.63-1.2,4,1.77a4.58,4.58,0,0,0,4.63-1s0.5-.1,1.13,1.26a87.15,87.15,0,0,0,9.5.61,10.86,10.86,0,0,1,1.75,4.5,9.9,9.9,0,0,0,3.88,1.13s-1.87,2.13,2.38,2.25c0,0-1.5,1.88.63,2,0,0-1.62,2.75,2.13,3.63,0,0,3.88,3.5,0,3.25,0,0-2.75,0-3.25-1.12,0,0-1.37.13-.25,2.25,0,0,2,5.13-2.12,2.38a5.49,5.49,0,0,0-1.87-.25s-3-2.12-3.87-1.62c0,0-.87.38-0.75,2,0,0-1.25,2.38-1.87-.5,0,0,1.63-3.12-2.5-.87l-0.12-1.87s-3-.87-3.12-6c0,0-6.62-1.5-6.25,1.13a8.52,8.52,0,0,0,3.88,1.5l3.63,5.38s-1.12,2.25.63,3.5l-3.87,7.38s-0.37,1.63,1.75,3.13c0,0-.87,2,0.5,2.5l-1,2.13a3.67,3.67,0,0,0,1.38,2.5s0,3.5-2.62,3.63c0,0-3,0-2,2.63,0,0-3,1.88-2.37,3.63,0,0-1.62,4.38-3.37,4.75a2.85,2.85,0,0,0,2.5,2.38s4.25,5.88-1.12,6H541.5s-1.75,3.5-1.12,4.5c0,0,3.63.38,4.38-.62,0,0,1.63-.49,2.25.69,0,0,1.63.43,1-1.57,0,0,4.13-5,7.63-5.87a23.44,23.44,0,0,1,0-3,4,4,0,0,0,1.88-3.75s4-.25,4-1.62l6.63,0.5s0.25,1.13,2-.62l4.13,0.38s1.88-1.37-1.25-1.5l-1.87-.25s-0.12-2.62,2.25-2.75c0,0,6.63-.87,7.88-2a3,3,0,0,1,3.13-.87s6.13-1.12,1-3.62c0,0-2.25-.87-0.87-1.62,0,0,1.88-2.62-4.75-.75l-2.75-.75s-0.37-.75,1-1.75v-2.75s2-1.87,3.38-1.75c0,0,3.5-3.37,3.88-4.37,0,0,10.63-3.25,12.88-5.37,0,0,.75-1.87,2.13-0.12a1.34,1.34,0,0,0,2.13,0,2.27,2.27,0,0,1,2.25.75s5,0.88,2.75-2c0,0-2.75-1.5-.12-2.87,0,0,1.5-2.25.13-2.5,0,0,8.25-3.87,11.88-3.5,0,0-1.25-1.37,0-4,0,0,2.38-1.25,1.63-3.62,0,0,3.38-2.62,2.75-4.75l6.88-3.25V535s8.13-2,0-4l-1.75-1.12s0-7.75-10,1.13l-3.37,4.25s-4.5-2-5.12-1.5c0,0-3.75,2.63-7-2.87,0,0-1.37-1.75-6-.62,0,0-3-1.25-2.75-2.87,0,0-12-2.12-16.37-.87L565,508.5l-13-59.62-1.5,3.75-6.75-3.5s-2.5,3.5-4,3.25c0,0-.87-2,1.88-4.5,0,0-.46-6.79-5.29-6.46,0,0,.58-5.58-3.83-1.5h-4.08s-0.08-3-4.83-2.67c0,0,1.33-5-3.42-6.08,0,0-6.93-6.42-7.8-7.67,0,0-3.62-12.25-2.78-14.83a4.42,4.42,0,0,0-3-3.08s-0.67-3.17.33-3.92c0,0,1.17-2.58-.17-2.75,0,0,1.42-2.5,2.58-2.75,0,0,1.42-1.25-1.83-1.42a11.24,11.24,0,0,1-2.33-1.5s-0.92-2.33-4-.83a18.87,18.87,0,0,1-4.08.58,3.68,3.68,0,0,1-3.58.33s-4.67-1.92-5-2.67c0,0-2.5.5-2.83,1.92l-8.33-.58s-0.25-4.42-5.08-3.83c0,0-2.67.17-2.75-.5,0,0-.5-1.58-2.92-0.75,0,0-1.75-2.33-4.08-1.42,0,0-1.92-2.17-3.67-1.33a9.47,9.47,0,0,1-1.58-1l-3.58-.08s0.5-4.58-3.83-4.83a4.83,4.83,0,0,1-2.75-1.58Z" transform="translate(-27.37 -231.75)" style="fill:#16a54a"/><path d="M552,448.88s1-11.29,8.13-3.75c0,0,1,1.13.92-1.12,0,0-.71-3.71,1.75-4,0,0,1.46-1.62-1.75-1.5,0,0-.83,2-2.21.75,0,0-1.83-5.63,1.67-7.81,0,0-.31-7.06-2.25-6.75,0,0-6.62-5.5-5.5-10.12,0,0-.44-2.94-2.19-2.81,0,0-3.37-1.44-.87-3.12a61.85,61.85,0,0,1-4.69-5.81s-5.81-1.56-5.75-3.62c0,0-3.31-.56-3.94-1.06,0,0,1.5-3.44,3.81-4.75,0,0,7-12.62,11.54-14.37,0,0-1.08-3.58,1.67-5,0,0,.33-6.08,2-7.42l-0.25-4.75,3.08-.33s0.17-4.24-1.92-2.7c0,0-2.33,2.62-2.17-1.63,0,0-.33-4-1.33-3.83,0,0-2.75-4.83-1.58-6.25,0,0-2.08-3.17-4-2.33a3.53,3.53,0,0,1-1.67-4.58s-0.5-6.42-6.17-4.5c0,0-2.83-2.83-5.08-1.66,0,0-3.83-1.83-2.33-4.58h-3.25a2.67,2.67,0,0,1,.58-2.33s0.67-7.33,3.08-8.42c0,0,1.17-.33,1-2.67,0,0,3.33-2.25,3.33-3.42,0,0-1.17-2.25-2.25-.67L531.33,315s-1.67,1-1.17-2.08c0,0,.33-3.58,2.08-3.92,0,0,.67-3.83-0.92-3.83,0,0-1.08-1.75,1.17-3.42,0,0-6.75.75-3.75-4.25,0,0-1.67-1.17.58-1.58,0,0,.63-1.83-1.73-0.75,0,0-3.27-1.25-4.52.75,0,0-3.5,1.17-1.67-2.25,0,0,3-2.75,1.17-4.5,0,0-.75-2.67.75-3,0,0,1.33-2.42-.58-2.67,0,0-1.17-1,1.33-2.08,0,0,0-2-.83-1.75,0,0-1-2.92.5-3.83,0,0-.75-4.5-2-4l-0.86-4.58s-4.13-5.42.78-5l2.25,0.58s2.42-2.83,5.83-1.92c0,0,5.58,1.17,7-1.08,0,0,.42-0.5,3.42,0a2.64,2.64,0,0,0,2.17-.92,6,6,0,0,1,4.83-.42s0.5,1.27,2.08-.32c0,0-.5-2.51,2.67-1l3.67,0.83s1,2.08,2.5,1.5l3.92-.33s1.25,2.67,2.33,2.67c0,0,.5.5-0.58,2.67a2.06,2.06,0,0,0,2.5-.5s3.17-2.92,3.75-.42a10.8,10.8,0,0,1-2.58,2.33s-1.25,1.92,2.58,1c0,0,3.58-1,4.08-.42,0,0,.67,1.58-1.5,3.67,0,0,0,3,2,1.25,0,0,2.17-.42.5,1.58,0,0,.42,1.83,2.25,0,0,0,6.58-2.58,8.17-2.08,0,0,2.83,1.5,2.83,2.75,0,0,1.58.75,1.92-1.5,0,0-.08-3.08,1.42-0.25a3.38,3.38,0,0,1-.5,4.5s-1.08,2.5,2.17,3.25c0,0,.33,3.08,1.17,3.33,0,0-.08,1.5-2.92,2.17a1.86,1.86,0,0,0,2.08,2l4.25,5.17v3s1.67,0.25.75,3.25c0,0-2.25,4.67,1.75,3.58,0,0,1.92-.92,2.58.83a1,1,0,0,0,1.25-.58s1.08-1.42,1.42.92c0,0,.17,3.33-1.83,3.25,0,0-.67,1.75,2.33,1.42,0,0,3.5-.25,3.92.58,0,0,1.83,1.42,1.58-.67L608,305.5s-0.92-2.08,1.58-2c0,0,2.58.33,2.92-1.42a1.48,1.48,0,0,1,2.5.17l1.75,0.92s0.08,3.58,1.25,3.08c0,0,1.17,0,1.08-2,0,0-.65.1,1.5,1.33a29.43,29.43,0,0,0,2.75,1.08s0.67,1.17,1.17.25c0,0-2.25-3.92,2-2.25,0,0,5.83-2,3.92-6.84,0,0-2.92-4.33,1.75-4.08,0,0-1.92-3.58-.83-3.75l-0.58-6.33s-2.58-1.5-.83-3.75c0,0,.08-2.58-1.17-2.17,0,0-.83-1.67,1.92-1l1.5,0.58v-4.67s-3.92-1.08-.58-3l0.58-3.67a1.42,1.42,0,0,1,1.67,1.25,1.66,1.66,0,1,1,0,.25s0.42,11.38,5.67,14.88a30.59,30.59,0,0,1,8,14.13s-0.75,10.25,6.63,9.88c0,0,10.38,3,7.25,14.25,0,0,.75,12.5,10.38,14.63,0,0,12.88,5.75,7.5,12.38,0,0-2.5,2.63-8.12-2.5,0,0-4.5-2.25-4.25,3.63,0,0,0,6.75-8,4.75,0,0-16.37-.75-3.5,16,0,0,6.38,4.25,4.5,10,0,0-2.37,7.13,1.5,8l10.25,0.5s3.88,4.75,4.5,6.88c0,0-5.75,5-.37,7.63,0,0,16-1.5,17.5-.87l6.25-8.12s1.38-1.5,2.63,1.63a4,4,0,0,0,5.25-.37l59.13-31.42,3,5s-0.83,1.5-3.17,1c0,0-.83,2.33-2.67,1.83,0,0-.25,8.08-6.75,6.5,0,0,1,1.5-1,1.58l-0.12,1.38s4.25,1.92.08,3.33c0,0,5,4.5,1.13,4.46,0,0,1.21,2.08-1.92,2.21,0,0-.62,1.31.06,2a1.09,1.09,0,0,1-.87,1.69s-1.75,1.69-.19,2.75a1.45,1.45,0,0,1-1.31,1.5s-3.12,2.5-2.87,3.25c0,0-2.94-.56-3,0.56,0,0-2.37,2.38-4.37,1.06,0,0-5.81,4.88-5.62,5.88,0,0-2.5-.06-2.94-1.25a1.71,1.71,0,0,0-1.94.19s-1.52,2-3.27,1.67c0,0-2.21.5-2.37,1.38,0,0-3.17,1.88-4.33.25,0,0,1,2.5-1.67,2.63,0,0-.5,2.54-3.92,1.38,0,0-2,.71-1.46,1.83,0,0-1,1.25-2.71,0,0,0-4.29,1.13-4.25,2.25,0,0-10.08,1.54-12.17,4.25,0,0-5.17,1.63-4.83,2.54,0,0-4.58-2.42-2.75.83,0,0,3.17,1.92-.67,2.08,0,0,.17,2.5-1.92,2.83,0,0,1.5,2.42-1,2.92,0,0,.42,6,1.42,6.92a2.07,2.07,0,0,1-2,2.83,9.84,9.84,0,0,1-5.75,2.25L677,448.88s3.83,3.79-1.83,3.71a1,1,0,0,0-.58,1.67s0.83,1-1.25,1.92c0,0,1.33,2.33-1.42,2.58,0,0,1.42,3.75-1.5,3.75,0,0,1.58,6.37-2.79,6.62,0,0-1,18.38-11,29.13a24.06,24.06,0,0,0,10.88-16.37,3.47,3.47,0,0,1,2.25-.25,22.59,22.59,0,0,1,9.88-20.75,6.34,6.34,0,0,1,3.63-4.87,24.08,24.08,0,0,1,7.38-7.75s5.88-12.75,25.13-11.62a24.63,24.63,0,0,1,5.5,2.5s1.25,0.63-4.37.75c0,0-1,1.38,3.63,1.25,0,0,3,1.25.63,2.75,0,0,3.75.38-.12,2l-2.37,3s2,3-.87,3.25c0,0,1.75,2.75-7.75,3.88l-9.62,3.25s-3.37,1-3.37,2l-0.12,2.38-12.25,4,2.88,8.13s-8.75,2-9.12,5c0,0-6,15.25-6,15.83l1.25,6.42s-5.83,9.08-5,10.5l1.08,3.17s4.5,3.5-3.42,2.67L661,526.75l-30.82,9.77L630.5,535s7.75-2.08,0-4l-1.75-1.12s0-7.58-10,1.13l-3.37,4.25s-3.75-1.62-5.12-1.5c0,0-3.25,3.17-7-2.87,0,0-1.71-1.54-6-.62,0,0-3.17-1-2.75-2.87,0,0-10-1.62-16.37-.87L565,508.5Z" transform="translate(-27.37 -231.75)" style="fill:#46a3db"/><path d="M633.83,267.25s0.17,0.75,3,0c0,0,3.76,5,1,6.42,0,0-.74,3.82,1,1.21,0,0,.2-3.17,2.16-2.29a1,1,0,0,1,.42,1.38s0.83,1.5,2.25-.21c0,0,2.38-1.83,2.33,1.42,0,0,.5,2.58,2.54,1.67,0,0,5.21,0,2,3.33a13.54,13.54,0,0,1-2.33.33s-1.33.54,0.25,1c0,0,3.12.42,3.67-.79,0,0,2.67-2.58,3,.38a2.17,2.17,0,0,1-2.08,1.58s-1.25,1.58.79,1.54c0,0,1.54-.08,1.83-0.75,0,0,4.87-2.12,2.21,3.29,0,0-2.37,1.83-3.33,1.92,0,0-1.83,1,.42,1.25,0,0,4.38-.08,5-2.25,0,0,3.75-2.46,3.42,1.33,0,0,3.88,1.5.33,2.46,0,0-3.5.21-3.5,2.17,0,0,1.21,1.37,4.75-1.46,0,0,1.54-2.33,3.13.75,0,0,4,.17.63,2.38,0,0-2,1.58.75,1.21a1.59,1.59,0,0,1,1.92.38s1.08,1.29.92-.46c0,0-.17-3.46,1.58-1.17l0.33,1.42s4.38-.21,1.25,1.75c0,0-.67,2.88-2.5,2.42,0,0-2,.42.88,1.21,0,0,1.75.17,1.88-.71a0.79,0.79,0,0,1,1.08.72s2.79-.64,1.29.86a3.73,3.73,0,0,1-2.17.67s-1.08,1.17.38,1.17c0,0,4.46.46,5-1.33,0,0,3.79-3,4,2.83,0,0,1.83,2.29-1.12,2.88,0,0-1.46.42,0.42,0.83,0,0,3.46.87,0.21,2.33,0,0-1.17,1.5.54,2.13,0,0,3.58.33,1.46,1.92,0,0-1.71,1.33.46,1.67l2.92,0.21s0.58-1.17,2.25,1.17c0,0,3.25,3.17,5.17,0,0,0,.67-4.17,4.17-2.58,0,0,3.83,1.67-.83,2.33,0,0-1.92,2,.92,1.66l1.4-1.41s1-.61,1.27,1.44v2.64s1.75-3.25.83,0c0,0,.17,4.17.83,4.58,0,0,.25,2.42,1.25,2.58,0,0,1.33-1.83,2.67-1.42,0,0,.25-3.25,1.5-1.5,0,0-.33,3.08,1.42,2.92,0,0,.75-2.67,0-4.25,0,0,0-3.08,1.83-.83,0,0,.33,1.92,1.42,1.5,0,0,1.17-1.25.42-2.33,0,0,.83-1.33,1.75.33,0,0,.5,3.83,1.33,1.92v-2.67s1.58-2.42,2.42-1.92c0,0,1.25.75,0,3.17,0,0,.42,3.08,1.58,1.75V324.5s2.5-3.25,4-1.08c0,0-1,2.5,1.17,2.42l2.17,0.5,1.75-2.08s-0.58-3.33,1.83-2.17a4.88,4.88,0,0,1,.58,2.42,2,2,0,0,0,1.92-.17s2-1.75,1.67.58c0,0-.33,1.5.92,1.08,0,0,1.5,0-1.75,2.42,0,0-3.83,4.92-3.83,6.25,0,0-2.5,3.08-5.5,2.17,0,0-2.83,3,.08,3.25l0.83-2s4.33,0.08,5.75-2.58c0,0,1.83-2.17,1.75,1.17,0,0-.17,2.83-2.42,3,0,0-2.75.17-2.83,2.08,0,0-1.58,1.83-2.67,1.25,0,0-4.42,9.25-8.25,8.83,0,0-2.75.5-5.92-2.42,0,0-3.08-1.42-3.42.33,0,0,.58,2.47,2.42,1.42a16.67,16.67,0,0,0,3.42,1.58s2.25,1.33,3.58,1.25c0,0,.92.92-1.33,1.75,0,0-1.08,4.17,3.17,2.58,0,0,1.92-2.25,2.58-1.75,0,0-.83-3.42,1.33-2.67,0,0,1.17-2.25,2-1.67l-1.92-.92s4.67-8.17,11.58-12.17c0,0-.39-4.75,1.68-5.67,0,0,4.18-3.67,0-1.83,0,0-5.26.58,0.24-1.83l3.58-1.75s3,2.17,3.75,2.92c0,0,3.17.08,0.67,3.25,0,0-2.17,2.58.58,2.92,0,0,2.33.67,2.08-.92,0,0,.08-1.67,1,1.25,0,0,3.17,3.67,1.83-1.5a2.5,2.5,0,0,0-1.5-2.08s-0.93-1.5,1.83-1.42c0,0,2.34-2.83,3.25.33a0.93,0.93,0,0,0,1.67.5s-0.58-1.67,1.33.33c0,0,1.92,1.5,2.25-.92,0,0,1.5-1,1.83,1.25,0,0,3,3.5,0,4.75,0,0-1.92,1.33,1.83,1.08,0,0,1.75,1.58-.42,2.42,0,0-1.33,1.33,1.08,1.25,0,0,3.75-1.92,1,1.42,0,0-1.67,1.25.08,1.5,0,0,2.58.42-1.25,1.08,0,0-2.25,2.33,1.17,1.33,0,0,1.92-1.25,2.83-1,0,0,.42-2.5,1.42-0.67,0,0,.42,1.5,1.58.25,0,0,2.83,1.42-1,2.42l-3.25.92s0.58,0.42,3.67.5c0,0,4.08-2.5,4.08-.58,0,0,.17,4.25-1.92,2.08,0,0,1.58,4.75-.17,5,0,0,0,2.92-1.67,3.08,0,0-.08,2.67-1.08,3a2.33,2.33,0,0,0,.5,1.67,1.76,1.76,0,0,1-1.75,1.67l-1.67,1.17-3-5-59.13,31.42s-2,3-5.25.38c0,0-1.66-2.58-2.35-1.83l-6.52,8.33s-6.33-.12-17.5.88c0,0-5.75-2.17.38-7.62,0,0-3.38-5.54-4.5-6.87l-10.25-.5s-4-.17-1.5-8c0,0,1.79-4.88-4.5-10,0,0-13.58-15.71,3.5-16,0,0,7.21,2.12,8-4.75,0,0-.79-5.92,4.25-3.62,0,0,5.58,4.75,8.13,2.5,0,0,5.58-6-7.5-12.37,0,0-9.12-.88-10.37-14.62,0,0,3.33-10.17-7.25-14.25,0,0-7.5.75-6.62-9.87a33.44,33.44,0,0,0-8-14.12S634.17,280,633.83,267.25Z" transform="translate(-27.37 -231.75)" style="fill:#2a347e"/><path d="M696.88,463.63l-12.25,4,2.88,8.13s4.5,1,6,4.75l8,19.25s8,6.5,10.5,9.5l9.25-3.25a3.52,3.52,0,0,1,4.5-3.5s6-9.25,9.25-13c0,0-.5-3.62,2.25-3.06,0,0,3.25-4.81,3.25,0,0,0-5.75,4.31-5.25,7.81,0,0,.5,1.5,1.75,0,0,0,4.5-3.52,7.75-3.26,0,0,6.92-4.78,7.17-1.62a2,2,0,0,1-2,.46s-4.5,1.71-4.33,5.13c0,0-.83,3.25-2.79.75a8.78,8.78,0,0,1-.17-2s-1.12-1.37-1.46.42L741,495.88s-3.46,1.17-3.08,2.67c0,0-3.08-.75-1.92,2.25,0,0-3.54,6.67-5.67,7A1.4,1.4,0,0,0,730,510s0.42,1-.92.92c0,0-3.54,2.38-2.87,4.71,0,0-.5,2.67,1.46.5,0,0,.46-3.38,1.88-3,0,0-3.25,5.46.17,7.46,0,0,.46.54,0.21,2.46,0,0-.75,3.42,1.75,1.46,0,0,2.21-2.46,3,.08,0,0-1.29,1.87,1.17,2,0,0,1.42,4.38,3,1,0,0,.54-1.33,1.46-1l1-2.62s1,0,.83-1.37a1.38,1.38,0,0,1,1.75.83s1.88,0.13.67-2.5a13.62,13.62,0,0,1,.54-2.25c0.13,0-1.92-1.33.63-1,0,0,2.46.83,0.83-1.58,0,0-1.08-5.71.83-6.92,0,0,3,.46,2.08-1.08L748.17,507s-0.62-1.42,1.17-1.37c0,0,1.71-.21.63-2.08,0,0-.17-0.79,1.63.17,0,0,1.38.17,1.42-1a1.27,1.27,0,0,1,1.46.25s1.67-.54,0-2,0.08-.21.08-0.21,2.92-2.42,3-3.83c0,0,2.71,1,2-.79,0,0,4.54.33,2-1.83,0,0,4-.54,1.79-2.54,0,0,3.33,1.33,3.33-1.25,0,0,2-2,2.92-1.83v-2.58s3.58-3.17,3.5-3.75c0,0,2.67.17,1.83-2,0,0-.38-1.15.41-1.44a1.24,1.24,0,0,1,.42-0.06c1.33,0,1.83.67,1.83-1a1.79,1.79,0,0,0,0-2.25s-4.58,1.92-4.67,3.67a1.11,1.11,0,0,1-1.67-.83s1.58-2.58-1.67-2.33c0,0-2.33.75-3.08-.17a1.84,1.84,0,0,0-2.25,1.42s-2.92,3-4.75,3L757,478.67a1.25,1.25,0,0,1-1.42,1.25s-0.58-2.83-3.33.25c0,0-2.75,1.67-4.42.75,0,0,.33-1.58-2.33.92,0,0-4.67,1.92-5.42.42l2.25-3.25s-0.42-2.25-4-.17c0,0-2.83.17-3.42,1.08a2.19,2.19,0,0,1-1.67-2.25s-1.92-1.5-2.75-1.33c0,0-.33-3.92-1.67-2.08v1.92s-2.5-3.42-1.5-4.17c0,0-2.5-1.08-.75-2.5,0,0-.08-2.75-2.42-1.17,0,0-1.25,2.58-2.83,1.83,0,0-7.25,4.67-2.42-.5,0,0,3.42-3.42,2.25-4,0,0,5.25-3.5-2.67-3.92,0,0,9.37-3.54,2.42-5,0,0-1.17.83-1.08,1.67,0,0-1.92.33-1.17-1.67,0,0-.75-1.42-2.25,2.33l-3.33,4.17s-0.58-3.42-3-2.5c0,0-.33-2.17-2.75-0.92,0,0-5.15,0-4.95-1.26,0,0-5.72,1.18-5.38,2.68Z" transform="translate(-27.37 -231.75)" style="fill:#2a347e"/><path d="M734.44,464.44s-1.69.38-.87,2.88c0,0-1.94-.25-1,3.31a1.81,1.81,0,0,0-.69,2.75s1.19,1.44,2.56-1.5c0,0,1.81-.81,1.94.63,0,0-.19,3.13.94,4.06,0,0,2.69-1.06,3-2.37,0,0,.94-0.62.94,1.06,0,0-1.44,1.5,2.25.81,0,0,.63-1.5,2.44-0.44,0,0,1.44,1.81,2,.06,0,0-2.81-2.37-.37-2.5,0,0,1.94-2.12,2.31.38v2.5a10.74,10.74,0,0,1,4.13.5s1.5-2.37,3.5-2.06c0,0-3.62-4.5,0-6.56,0,0,4.25-1.56,2.5-3.5,0,0-1.25-1.62-5.5,2.06a4,4,0,0,1-3,1.19s-1.06-.81-0.69,1.25a1.48,1.48,0,0,1-1.31.25,1.05,1.05,0,0,0-1.5,1.13,6,6,0,0,1-2.87.3s-0.94,0,.13,1.38c0,0,.25,1.25-1.5-0.31,0,0-1.44-1.87-3.12-1.06l-0.06,1.88-1.62-.12a2.93,2.93,0,0,0-3.56-2.94s1.88-2.94-.25-3.19C735.13,466.25,735.69,464.25,734.44,464.44Z" transform="translate(-27.37 -231.75)" style="fill:#2a347e"/><path d="M783.25,357.63a8.93,8.93,0,0,1-2.75,1.17s-1.25.45,0,2.7c0,0-1.12,1.25-1.75,0,0,0-.74-4.14-1.5.56,0,0-.63,4.82-2.88,5.44,0,0-1.87,1.75,0,3.25,0,0,.71,2.5-0.83,2.63,0,0-.92,1.75.58,1.75,0,0,1.78,2.88-.86,4.88,0,0-1.64,2.25,1.36,2.13,0,0-1,2.63.5,3.13,0,0,.38,4.63-0.75,5.25,0,0,2.38,2.13.63,2.63,0,0-.87,2.75,1,4,0,0-1.12,8.38,1.25,8.88,0,0,5-2.5,3.5,2.38,0,0-2.37,1.38-4.37.75,0,0-1.37-.62-0.5,2.88,0,0,1.25,6-.37,6.5,0,0-1.12.13-1.12-2.12,0,0-2.75.25-2.75,5.75,0,0-.37,2,1.88,0,0,0,4.25-4,5.38-4h-1.12l3.63-.25s1.13-2.62,2.5-.87c0,0-4.27,5.78-6.14,6v2.49s-3.36,4.5-3,5.38h3s-0.49,2.25.89,2.63c0,0-.12,3.63,2.88.75a3.13,3.13,0,0,1,2.38-1.5s3.13-2.75,2.38-3.62c0,0,2.5,1.88,4-1.62,0,0,.5-1.75,2.75-0.37l1-1.5,2.25,0.5s-0.5-4,1.75-1.75c0,0,.75,2.38,1.25.25l3.63-.75s2.38-1.75,1.75-2.62a2.86,2.86,0,0,1,2.5-1.37s-0.12-1.12,1.88-.62v-2.62l1.5-.37s-1.12-3.75.88-4a4.79,4.79,0,0,1,0,3.63l1.38,0.75,1.88-3.75a5.49,5.49,0,0,0,1.38,1.75s0.88-2,1.88.63c0,0-1,3.25,1.5,4.63,0,0,.13,4.13-2.87,5.13,0,0-1.87,3.63.5,3.88a2.54,2.54,0,0,0,1.5-1.62,2.24,2.24,0,0,0,2.25-1.87l2.88-.75s2.88-4.75,0-6.37l1.88-1.62V414s1.63-2.5,0-3.5a2,2,0,0,1,1.13-2.19s3.63-3.69.38-4.25a2.87,2.87,0,0,1,3,.94s0.25,1.75,3.06,1.94c0,0-.62,10.88,3.25,12.38a3.92,3.92,0,0,0,.81-3.62s1.38-2.87-.12-3.44c0,0,.25-3.31,1.88-0.87,0,0,1.75,2.06,1.56,3.63,0,0,.5,2.19,1.81,1.88,0,0-.19,2.44,1.38,1.94,0,0-.79-4.23,3.71-3.73,0,0,1.33-1.17-.42-4.75,0,0-2.75-1.67-1.42-4.42,0,0-.92-8.5-5.33-8.25,0,0-2,.58-0.42,2.25,0,0,3.67,5.25.08,5.75,0,0-3.17.5-2.17-4,0,0,1.5-3.92-1.17-4.33,0,0,1.67-4.25.17-3.67,0,0-2.58,1.33-2,3.58,0,0,.25,6.83-2.17,6.42,0,0-2,1.25-1.75-4.92,0,0-.58-1.58-2.08.5,0,0-3.92,3.42-.75-0.92,0,0,3.92-4.67,2.75-5.92,0,0,3-5.67-.42-1.92,0,0-4.17,2.25-3.75,3.92,0,0-1.67,1.25-1.92-1.58,0,0,.83-3.92-1.67-2.67a2.22,2.22,0,0,0-2.42-1s2.75-5.17.67-5.42c0,0,2.67-2.83-2.42-2.08,0,0-2.33.42-3.17-.42a1.5,1.5,0,0,0-1.25,2s-3.17-1.17-2.08,2.83c0,0-.67,1.83-1.17-1.17,0,0-.67-4.58-2.67-2,0,0-.33,2.33.25,2.92,0,0-2.33.25-1.33,2.33,0,0-3.58-1.42-1,4.83,0,0-.83,1.42-1.92-0.83,0,0-.58-3.25.5-4.75,0,0-2.17-2.92-3.08-.58,0,0-.76,2.42-1.84,2.47a0.78,0.78,0,0,1-.33-0.06c-1.25-.5-4.83-1.58-4.67,1.17,0,0,1.83-4.58-1.58-2.67,0,0,1.5-5.17,3.83-5.67,0,0,.25-2.67-3.58-0.08a2.25,2.25,0,0,0-2.17,2s-1.33-3.92-2.33-3.17c0,0-1.25,1.25-.67,2.42,0,0-2.17,1.17-.17,4.08,0,0,.17,1.25-0.83,1.33,0,0-.58,4.5-0.25,5.92,0,0-1.17,1-1.33-.5,0,0,.58-1.67-0.58-1.67,0,0,0-7.17-1.25-7.33a3,3,0,0,1,.75-2.33s-1-4.58-2.17-4.67c0,0,1.5-4.58-.33-5.08,0,0,4.42-.75,1.83-3.25,0,0-.67-2.25.58-3.25,0,0-.67-2.58-2.58-0.42,0,0-3.33-.42-1.17-1.92,0,0,1.42-1.58,2.58-1.33C783.58,362.83,782.67,357.83,783.25,357.63Z" transform="translate(-27.37 -231.75)" style="fill:#2a347e"/><path d="M770.25,449.25s-0.75-3.25-2.62.75c0,0,1.25,2.38.75,3.25,0,0-2.62-.28-0.75,3.24,0,0-1.62,1.64.25,4.39a1.28,1.28,0,0,1-.92,2s-1.8.25,0,3c0,0-1.17,5.88,1.25,6.63,0,0,4.55.38,5.42,1.13,0,0,2.75,3.5,1.88-1.12,0,0-.19-0.37,2.41-0.37a3.7,3.7,0,0,0,2-3.75s3,2.13,2.13-2.37c0,0-.37-2.09,2-1.92v-1.33s6.88-4.23,0-1.3c0,0-1.75.3-1-1.7,0,0-2.87-2.62-3.87.25,0,0-3.92.25-2.52,3.13,0,0-.1,2.25-1,2.75,0,0-.19,2.41,2.28-1.79,0,0,1.09.67,0.59,2.54,0,0-.37,1.38-1.25,1.38,0,0-1.17.5-.65,2.32h-4.1a2.87,2.87,0,0,1,1.25-3.2s-2.12-1-.5-2.25c0,0,2.88-.6,1-3.43,0,0-.37-6.32.13-7.2l-1.25-1.62s0.2-4-1.52-2C771.6,450.63,770,451.63,770.25,449.25Z" transform="translate(-27.37 -231.75)" style="fill:#2a347e"/><circle cx="66.44" cy="221.23" r="1.98" style="fill:#feeb00"/><circle cx="69.82" cy="223.21" r="1.98" style="fill:#feeb00"/><circle cx="73.15" cy="223.21" r="1.98" style="fill:#feeb00"/><circle cx="98.26" cy="207.25" r="1.98" style="fill:#feeb00"/><circle cx="169.3" cy="224.19" r="1.98" style="fill:#feeb00"/><circle cx="181.38" cy="182.5" r="1.98" style="fill:#feeb00"/><circle cx="244.13" cy="214.25" r="1.98" style="fill:#feeb00"/><circle cx="536.51" cy="283.88" r="1.98" style="fill:#feeb00"/><circle cx="533.53" cy="352.55" r="1.98" style="fill:#feeb00"/><circle cx="542.46" cy="342.5" r="1.98" style="fill:#feeb00"/><circle cx="550.96" cy="334.88" r="1.98" style="fill:#feeb00"/><circle cx="554.92" cy="337.63" r="1.98" style="fill:#feeb00"/><circle cx="538.48" cy="344.48" r="1.98" style="fill:#feeb00"/><circle cx="589.9" cy="302.55" r="1.98" style="fill:#feeb00"/><circle cx="726.38" cy="268.77" r="1.98" style="fill:#feeb00"/><circle cx="716.63" cy="250.02" r="1.98" style="fill:#feeb00"/><circle cx="610.63" cy="294.5" r="1.98" style="fill:#feeb00"/><circle cx="627.38" cy="266.53" r="1.98" style="fill:#feeb00"/><circle cx="738.63" cy="246.06" r="1.98" style="fill:#feeb00"/><circle cx="685.38" cy="264.55" r="1.98" style="fill:#feeb00"/><circle cx="755.38" cy="230.13" r="1.98" style="fill:#feeb00"/>'
